pyproject.toml
[tool.poetry]
name = "vfn-rag"
version = "0.1.0"
description = "Overall package"
authors = ["Mostafa Farrag <mostafa.farrag@deltares.nl>"]
license = "Proprietary"
readme = "README.md"
keywords = ["llm", "generativeai", "chatbot", "rag", "natural language processing"]
repository = "https://github.com/Deltares-research/vfn-rag"

packages = [{ include = "vfn_rag", from = "src"}]

[tool.poetry.scripts]
vfn-rag = "vfn_rag.cli:cli"


[tool.poetry.urls]

[tool.poetry.dependencies]
python = ">=3.12,<4"
llama-index = "^0.12.36"
opencv-python = "^4.11.0.86"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
click = "^8.1.7"
azure-cosmos = "^4.5.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.llm.dependencies]
llama-index-llms-azure-openai = "^0.3.2"
llama-index-embeddings-azure-openai = "^0.3.5"
llama-index-vector-stores-azurecosmosnosql = "^1.3.0"


[tool.poetry.group.analysis.dependencies]
jupyterlab= "^4.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"
xdoctest = {extras = ["colors"], version = "^1.1.0"}
typeguard = "^4.3.0"
nbval = "^0.11.0"
coverage = {extras = ["toml"], version = "^7.6.2"}
safety = "^3.2.8"
pre-commit = "^4.0.1"



[tool.coverage.report]
show_missing = true
fail_under = 10

[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88
ensure_newline_before_comments=true
profile="black"


[tool.flake8]
ignore = ["E501"]
max-line-length = 88

[tool.black]
profile = "flake8"
line-length = 88
skip-string-normalization = true  # Avoid unnecessary quote changes


[tool.pytest.ini_options]
addopts = [ "--cov"]

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
